+++ lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_navigation_methods.rb
@@ -20,11 +20,15 @@ module OnlyofficeWebdriverWrapper
 
     # @return [String] url of current frame, or browser url if
     #   it is a root frame
-    def get_url
+    def current_url
       execute_javascript('return window.location.href')
     rescue Selenium::WebDriver::Error::NoSuchFrameError, Timeout::Error => e
       raise(e.class, "Browser is crushed or hangup with #{e}")
     end
+
+    alias get_url current_url
+    extend Gem::Deprecate
+    deprecate :get_url, 'current_url', 2069, 1
 
     # Refresh current page
     # @return [nil]
@@ -48,11 +52,11 @@ module OnlyofficeWebdriverWrapper
       begin
         @driver.execute_script('window.onbeforeunload = null') # off popup window
       rescue StandardError
-        Exception
+        StandardError
       end
       begin
         @driver.quit
-      rescue Exception => e
+      rescue StandardError => e
         OnlyofficeLoggerHelper.log("Some error happened on webdriver.quit #{e.backtrace}")
       end
       alert_confirm