+++ spec/onlyoffice_webdriver_wrapper/web_driver/firefox_spec.rb
@@ -2,8 +2,12 @@
 
 require 'spec_helper'
 
-describe 'Webdriver Firefox' do
-  let(:webdriver) { OnlyofficeWebdriverWrapper::WebDriver.new(:firefox) }
+describe OnlyofficeWebdriverWrapper::WebDriver, '#firefox' do
+  let(:webdriver) { described_class.new(:firefox) }
+
+  after do
+    webdriver.quit
+  end
 
   it 'Check That firefox start correctly' do
     expect { webdriver }.not_to raise_error
@@ -34,9 +38,5 @@ describe 'Webdriver Firefox' do
   it 'Check that firefox can download files by direct links' do
     result = webdriver.download('https://onlyoffice-webdriver-wrapper-spec-files.s3.amazonaws.com/0+(1).docx')
     expect(File).to exist(result)
-  end
-
-  after do
-    webdriver.quit
   end
 end